port clams-whisper to hugging face implementation
